src/infra/cryptography/bcrypt-adapter.ts
import bcrypt from 'bcrypt'
import { HashComparer } from '../../data/protocols/cryptography/hash-comparer'

import { Hasher } from '../../data/protocols/cryptography/hasher'

export class BCryptAdapter implements Hasher, HashComparer {
  constructor (private readonly salt: number) {}
  async hash (value: string): Promise<string> {
    const hash = await bcrypt.hash(value, this.salt)
    return await new Promise<string>(resolve => resolve(hash))
  }

  async compare (value: string, hash: string): Promise<boolean> {
    await bcrypt.compare(value, hash)
    return await new Promise<boolean>(resolve => resolve(true))
  }
}
